--- lib/components/search_bar.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-
-class SearchBar extends StatelessWidget {
-  const SearchBar({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.symmetric(vertical: 30),
-      padding: EdgeInsets.symmetric(horizontal: 30, vertical: 5),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(29.5),
-      ),
-      child: TextField(
-        decoration: InputDecoration(
-          hintText: "Поиск",
-          icon: SvgPicture.asset("icons/search.svg"),
-          border: InputBorder.none,
-        ),
-      ),
-    );
-  }
-}
