+++ lib/screens/settings/group_selection_screet.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class GroupSelectionScreen extends StatefulWidget {
-  GroupSelectionScreen({Key key}) : super(key: key);
+  GroupSelectionScreen({Key? key}) : super(key: key);
 
   @override
   _GroupSelectionScreenState createState() => _GroupSelectionScreenState();
